put log app start to the separate module
